Added tools for memory benchmark
Added `memtest.sh` which is a simple script that evaluates the memory
usage as a function of time of a process. The pipeline now includes a
`membenchmark` target in the mappers Makefile, which performs the memory
benchmark using only one of the invocations of the mapper, with one
million reads.